+++ provider.go
@@ -26,7 +26,8 @@ func Provider() *schema.Provider {
 	return &schema.Provider{
 		ConfigureFunc: providerConfigure,
 		DataSourcesMap: map[string]*schema.Resource{
-			"proxmox_virtual_environment_version": dataSourceVirtualEnvironmentVersion(),
+			"proxmox_virtual_environment_access_groups": dataSourceVirtualEnvironmentAccessGroups(),
+			"proxmox_virtual_environment_version":       dataSourceVirtualEnvironmentVersion(),
 		},
 		ResourcesMap: map[string]*schema.Resource{},
 		Schema: map[string]*schema.Schema{