+++ proxmox/virtual_environment_version.go
@@ -5,11 +5,7 @@
 package proxmox
 
 import (
-	"bytes"
-	"encoding/json"
 	"errors"
-	"fmt"
-	"net/http"
 )
 
 // VirtualEnvironmentVersionResponseBody contains the body from a version response.
@@ -27,35 +23,11 @@ type VirtualEnvironmentVersionResponseData struct {
 
 // Version retrieves the version information.
 func (c *VirtualEnvironmentClient) Version() (*VirtualEnvironmentVersionResponseData, error) {
-	req, err := http.NewRequest("GET", fmt.Sprintf("%s/%s/version", c.Endpoint, basePathJSONAPI), new(bytes.Buffer))
-
-	if err != nil {
-		return nil, errors.New("Failed to create version information request")
-	}
-
-	err = c.AuthenticateRequest(req)
+	resBody := &VirtualEnvironmentVersionResponseBody{}
+	err := c.DoRequest(hmGET, "version", nil, resBody)
 
 	if err != nil {
 		return nil, err
-	}
-
-	res, err := c.httpClient.Do(req)
-
-	if err != nil {
-		return nil, errors.New("Failed to perform version information request")
-	}
-
-	err = c.ValidateResponse(res)
-
-	if err != nil {
-		return nil, err
-	}
-
-	resBody := VirtualEnvironmentVersionResponseBody{}
-	err = json.NewDecoder(res.Body).Decode(&resBody)
-
-	if err != nil {
-		return nil, errors.New("Failed to decode version information response")
 	}
 
 	if resBody.Data == nil {