+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 FEATURES:
 
+* **New Data Source:** `proxmox_virtual_environment_access_groups`
 * **New Data Source:** `proxmox_virtual_environment_version`